Correção Final

+++ ValueObjects/ParamQueryCalculaJuros.cs
@@ -2,7 +2,7 @@ namespace DualAPI.ValueObjects
 {
     public class ParamQueryCalculaJuros
     {
-        public double ValorInicial { get; set; }
+        public decimal ValorInicial { get; set; }
         public int Meses { get; set; }
     }
 }
